hipMemcpy: Treat Dst==Src as a NOP
Although ROCm API ref says that Dst and Src should not overlap,
HIP seems to handle Dst == Src as a special (no-operation) case.

Intel GPU OpenCL driver seems to do also so for clEnqueueSVMMemcpy,
which make it pass, but Intel CPU OpenCL returns CL_​MEM_​COPY_​OVERLAP
like it should. To unify the behavior, let's convert the special case
to a maker here, so we can return an event.

Fixes #147.