Fix OpenCL emulation of host-mapped memory.
The problem was visible with the hipHostRegister test which
allocates a host buffer with malloc() and then calls hipHostRegister()
to register it and hipHostGetDevicePointer() to get the device-side
shadow buffer.  The copy was erroneously skipped in this case
since there was a false assumption that with the OpenCL BE we
allocate everything via clSVMAlloc().  The skipping happens in
CHIPBackend.cc:1610.

There should be a better logic that checks if the buffer was originally
allocated via SVM in this case instead so we can skip the shadow buf
synch safely.

Also renames

 virtual bool isAllocatedPtrUSM(void *Ptr) = 0;
 to
 virtual bool isAllocatedPtrMappedToVM(void *Ptr) = 0;

to avoid confusion.

USM = Unified Shared Memory, which is the LevelZero
memory model and the Intel OpenCL extension:
https://registry.khronos.org/OpenCL/extensions/intel/cl_intel_unified_shared_memory.html